+++ src/main/java/Jane/JaneException.java
@@ -1,0 +1,8 @@
+package jane;
+
+
+public class JaneException extends Exception {
+    public JaneException(String err) {
+        super(err);
+    }
+}
